pyproject.toml
[tool.poetry]
name = "keateka-backend"
version = "0.1.0"
description = "KeaTeka Cleaning Service Backend"
authors = ["James Wanjiku <james544@wanjiku@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12.6"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.10.2"}
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}  # Replaced python-jose with PyJWT
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = {extras = ["hiredis"], version = "^5.2.0"}
firebase-admin = "^6.6.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.10"
pytz = "^2024.2"
pytest = "^8.3.3"
pyhamcrest = "^2.1.0"
injector = "^0.22.0"
googlemaps = "^4.10.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
ruff = "^0.8.0"
pre-commit = "^4.0.1"
types-python-dateutil = "^2.9.0.20241003"
types-pytz = "^2024.2.0.20241003"
types-requests = "^2.32.0.20241016"
types-setuptools = "^69.0.0.20240125"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"
